backend/requirements.txt
aniso8601==8.1.1
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.1.1
colorlog==6.7.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.7.0
django-graphql-jwt==0.3.2
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
idna==2.10
inflection==0.5.1
iso8601==1.0.2
itypes==1.2.0
Jinja2==3.0.0
MarkupSafe==2.0.0
numpy==1.23.3
packaging==20.9
ply==3.11
promise==2.3
PyJWT==2.1.0
pyparsing==2.4.7
pytz==2021.1
pyxirr==0.7.2
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
Rx==1.6.1
scipy==1.9.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
xirr==0.1.8
